split health_check.cpp
