+++ ydb/core/health_check/health_check.cpp
@@ -1,42 +1,27 @@
+#include "merge_issues.h"
+#include "group_checker.h"
 #include "health_check.h"
+#include "health_check_structs.h"
+#include "health_check_utils.h"
 
-#include <yql/essentials/public/issue/yql_issue_message.h>
-#include <yql/essentials/public/issue/yql_issue.h>
-
-#include <ydb/library/actors/core/interconnect.h>
-#include <ydb/library/actors/core/hfunc.h>
-#include <ydb/library/actors/interconnect/interconnect.h>
-#include <library/cpp/digest/old_crc/crc.h>
-#include <library/cpp/protobuf/json/proto2json.h>
-#include <ydb/public/sdk/cpp/src/library/grpc/client/grpc_client_low.h>
-
-#include <util/random/shuffle.h>
-
-#include <ydb/core/base/hive.h>
 #include <ydb/core/base/path.h>
 #include <ydb/core/base/statestorage.h>
-#include <ydb/core/base/tablet_pipe.h>
-#include <ydb/core/cms/console/configs_dispatcher.h>
 #include <ydb/core/mon/mon.h>
-#include <ydb/core/base/nameservice.h>
 #include <ydb/core/blobstorage/base/blobstorage_events.h>
-#include <ydb/core/cms/console/console.h>
 #include <ydb/core/mind/tenant_slot_broker.h>
-#include <ydb/core/node_whiteboard/node_whiteboard.h>
-#include <ydb/core/tx/scheme_cache/scheme_cache.h>
 #include <ydb/core/tx/schemeshard/schemeshard.h>
 #include <ydb/core/util/proto_duration.h>
 #include <ydb/core/util/tuples.h>
 
-#include <ydb/core/protos/blobstorage_distributed_config.pb.h>
 #include <ydb/core/protos/config.pb.h>
-#include <ydb/core/sys_view/common/events.h>
-
-#include <ydb/public/api/grpc/ydb_monitoring_v1.grpc.pb.h>
-#include <regex>
-
+#include <ydb/library/actors/core/interconnect.h>
+#include <ydb/library/actors/interconnect/interconnect.h>
 #include <ydb/library/actors/wilson/wilson_span.h>
 #include <ydb/library/wilson_ids/wilson.h>
+#include <ydb/public/sdk/cpp/src/library/grpc/client/grpc_client_low.h>
+#include <library/cpp/protobuf/json/proto2json.h>
+
+#include <regex>
 
 static decltype(auto) make_vslot_tuple(const NKikimrBlobStorage::TVSlotId& id) {
     return std::make_tuple(id.GetNodeId(), id.GetPDiskId(), id.GetVSlotId());
@@ -131,510 +116,6 @@ public:
         , HealthCheckConfig(config)
     {}
 
-    using TGroupId = ui32;
-
-    enum class ETags {
-        None,
-        DBState,
-        StorageState,
-        PoolState,
-        GroupState,
-        VDiskState,
-        PDiskState,
-        NodeState,
-        VDiskSpace,
-        PDiskSpace,
-        ComputeState,
-        TabletState,
-        SystemTabletState,
-        OverloadState,
-        SyncState,
-        Uptime,
-        QuotaUsage,
-    };
-
-    enum ETimeoutTag {
-        TimeoutBSC,
-        TimeoutFinal,
-    };
-
-    struct TTenantInfo {
-        TString Name;
-        Ydb::Cms::GetDatabaseStatusResult::State State;
-    };
-
-    struct TNodeTabletState {
-        struct TTabletStateSettings {
-            TInstant AliveBarrier;
-            ui32 MaxRestartsPerPeriod; // per hour
-            ui32 MaxTabletIdsStored = 10;
-            bool ReportGoodTabletsIds = false;
-        };
-
-        enum class ETabletState {
-            Good,
-            Stopped,
-            RestartsTooOften,
-            Dead,
-        };
-
-        struct TNodeTabletStateCount {
-            NKikimrTabletBase::TTabletTypes::EType Type;
-            ETabletState State;
-            bool Leader;
-            int Count = 1;
-            TStackVec<TString> Identifiers;
-
-            static ETabletState GetState(const NKikimrHive::TTabletInfo& info, const TTabletStateSettings& settings) {
-                if (info.volatilestate() == NKikimrHive::TABLET_VOLATILE_STATE_STOPPED) {
-                    return ETabletState::Stopped;
-                }
-                ETabletState state = (info.restartsperperiod() >= settings.MaxRestartsPerPeriod) ? ETabletState::RestartsTooOften : ETabletState::Good;
-                if (info.volatilestate() == NKikimrHive::TABLET_VOLATILE_STATE_RUNNING) {
-                    return state;
-                }
-                if (info.tabletbootmode() != NKikimrHive::TABLET_BOOT_MODE_DEFAULT) {
-                    return state;
-                }
-                if (info.lastalivetimestamp() != 0 && TInstant::MilliSeconds(info.lastalivetimestamp()) < settings.AliveBarrier) {
-                    // Tablet is not alive for a long time
-                    // We should report it as dead unless it's just waiting to be created
-                    if (info.generation() == 0 && info.volatilestate() == NKikimrHive::TABLET_VOLATILE_STATE_BOOTING && !info.inwaitqueue()) {
-                        return state;
-                    }
-                    return ETabletState::Dead;
-                }
-                return state;
-
-            }
-
-            TNodeTabletStateCount(const NKikimrHive::TTabletInfo& info, const TTabletStateSettings& settings)
-                : Type(info.tablettype())
-                , State(GetState(info, settings))
-                , Leader(info.followerid() == 0)
-            {
-            }
-
-            bool operator ==(const TNodeTabletStateCount& o) const {
-                return State == o.State && Type == o.Type && Leader == o.Leader;
-            }
-        };
-
-        TStackVec<TNodeTabletStateCount> Count;
-
-        void AddTablet(const NKikimrHive::TTabletInfo& info, const TTabletStateSettings& settings) {
-            TNodeTabletStateCount tabletState(info, settings);
-            auto itCount = Find(Count, tabletState);
-            if (itCount != Count.end()) {
-                itCount->Count++;
-            } else {
-                Count.emplace_back(tabletState);
-                itCount = std::prev(Count.end());
-            }
-            if (itCount->State != ETabletState::Good || settings.ReportGoodTabletsIds) {
-                if (itCount->Identifiers.size() < settings.MaxTabletIdsStored) {
-                    TStringBuilder id;
-                    id << info.tabletid();
-                    if (info.followerid()) {
-                        id << '.' << info.followerid();
-                    }
-                    itCount->Identifiers.emplace_back(id);
-                }
-            }
-        }
-    };
-
-    struct TStoragePoolState {
-        TString Name;
-        THashSet<TGroupId> Groups;
-    };
-
-    struct TDatabaseState {
-        TTabletId HiveId = {};
-        TTabletId SchemeShardId = {};
-        TPathId ResourcePathId = {};
-        TVector<TNodeId> ComputeNodeIds;
-        THashSet<ui64> StoragePools;        // BSConfig case
-        THashSet<TString> StoragePoolNames; // no BSConfig case
-        THashMap<std::pair<TTabletId, TFollowerId>, const NKikimrHive::TTabletInfo*> MergedTabletState;
-        THashMap<TNodeId, TNodeTabletState> MergedNodeTabletState;
-        THashMap<TNodeId, ui32> NodeRestartsPerPeriod;
-        ui64 StorageQuota = 0;
-        ui64 StorageUsage = 0;
-        TMaybeServerlessComputeResourcesMode ServerlessComputeResourcesMode;
-        TNodeId MaxTimeDifferenceNodeId = 0;
-        TString Path;
-    };
-
-    struct TGroupState {
-        TString ErasureSpecies;
-        std::vector<const NKikimrSysView::TVSlotEntry*> VSlots;
-        ui32 Generation;
-        bool LayoutCorrect = true;
-    };
-
-    struct TSelfCheckResult {
-        struct TIssueRecord {
-            Ydb::Monitoring::IssueLog IssueLog;
-            ETags Tag;
-        };
-
-        Ydb::Monitoring::StatusFlag::Status OverallStatus = Ydb::Monitoring::StatusFlag::GREY;
-        TList<TIssueRecord> IssueRecords;
-        Ydb::Monitoring::Location Location;
-        int Level = 1;
-        TString Type;
-
-        static bool IsErrorStatus(Ydb::Monitoring::StatusFlag::Status status) {
-            return status != Ydb::Monitoring::StatusFlag::GREEN;
-        }
-
-        static TString crc16(const TString& data) {
-            return Sprintf("%04x", (ui32)::crc16(data.data(), data.size()));
-        }
-
-        static TString crc32(const TString& data) {
-            return Sprintf("%08x", (ui32)::crc32(data.data(), data.size()));
-        }
-
-        static TString GetIssueId(const Ydb::Monitoring::IssueLog& issueLog) {
-            const Ydb::Monitoring::Location& location(issueLog.location());
-            TStringStream id;
-            if (issueLog.status() != Ydb::Monitoring::StatusFlag::UNSPECIFIED) {
-                id << Ydb::Monitoring::StatusFlag_Status_Name(issueLog.status()) << '-';
-            }
-            id << crc16(issueLog.message());
-            if (location.database().name()) {
-                id << '-' << crc32(location.database().name());
-            }
-            if (location.storage().node().id()) {
-                id << '-' << location.storage().node().id();
-            } else {
-                if (location.storage().node().host()) {
-                    id << '-' << location.storage().node().host();
-                }
-                if (location.storage().node().port()) {
-                    id << '-' << location.storage().node().port();
-                }
-            }
-            if (!location.storage().pool().group().vdisk().id().empty()) {
-                id << '-' << location.storage().pool().group().vdisk().id()[0];
-            } else {
-                if (!location.storage().pool().group().id().empty()) {
-                    id << '-' << location.storage().pool().group().id()[0];
-                } else {
-                    if (location.storage().pool().name()) {
-                        id << '-' << crc32(location.storage().pool().name());
-                    }
-                }
-            }
-            if (!location.storage().pool().group().vdisk().pdisk().empty() && location.storage().pool().group().vdisk().pdisk()[0].id()) {
-                id << '-' << location.storage().pool().group().vdisk().pdisk()[0].id();
-            }
-            if (location.compute().node().id()) {
-                id << '-' << location.compute().node().id();
-            } else {
-                if (location.compute().node().host()) {
-                    id << '-' << location.compute().node().host();
-                }
-                if (location.compute().node().port()) {
-                    id << '-' << location.compute().node().port();
-                }
-            }
-            if (location.compute().pool().name()) {
-                id << '-' << location.compute().pool().name();
-            }
-            if (location.compute().tablet().type()) {
-                id << '-' << location.compute().tablet().type();
-            }
-            if (location.compute().schema().path()) {
-                id << '-' << crc32(location.compute().schema().path());
-            }
-            return id.Str();
-        }
-
-        void ReportStatus(Ydb::Monitoring::StatusFlag::Status status,
-                          const TString& message = {},
-                          ETags setTag = ETags::None,
-                          std::initializer_list<ETags> includeTags = {}) {
-            OverallStatus = MaxStatus(OverallStatus, status);
-            if (IsErrorStatus(status)) {
-                std::vector<TString> reason;
-                if (includeTags.size() != 0) {
-                    for (const TIssueRecord& record : IssueRecords) {
-                        for (const ETags& tag : includeTags) {
-                            if (record.Tag == tag) {
-                                reason.push_back(record.IssueLog.id());
-                                break;
-                            }
-                        }
-                    }
-                }
-                std::sort(reason.begin(), reason.end());
-                reason.erase(std::unique(reason.begin(), reason.end()), reason.end());
-                TIssueRecord& issueRecord(*IssueRecords.emplace(IssueRecords.begin()));
-                Ydb::Monitoring::IssueLog& issueLog(issueRecord.IssueLog);
-                issueLog.set_status(status);
-                issueLog.set_message(message);
-                if (Location.ByteSizeLong() > 0) {
-                    issueLog.mutable_location()->CopyFrom(Location);
-                }
-                issueLog.set_id(GetIssueId(issueLog));
-                if (Type) {
-                    issueLog.set_type(Type);
-                }
-                issueLog.set_level(Level);
-                if (!reason.empty()) {
-                    for (const TString& r : reason) {
-                        issueLog.add_reason(r);
-                    }
-                }
-                if (setTag != ETags::None) {
-                    issueRecord.Tag = setTag;
-                }
-            }
-        }
-
-        bool HasTags(std::initializer_list<ETags> tags) const {
-            for (const TIssueRecord& record : IssueRecords) {
-                for (const ETags tag : tags) {
-                    if (record.Tag == tag) {
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-
-        Ydb::Monitoring::StatusFlag::Status FindMaxStatus(std::initializer_list<ETags> tags) const {
-            Ydb::Monitoring::StatusFlag::Status status = Ydb::Monitoring::StatusFlag::GREY;
-            for (const TIssueRecord& record : IssueRecords) {
-                for (const ETags tag : tags) {
-                    if (record.Tag == tag) {
-                        status = MaxStatus(status, record.IssueLog.status());
-                    }
-                }
-            }
-            return status;
-        }
-
-        void ReportWithMaxChildStatus(const TString& message = {},
-                                        ETags setTag = ETags::None,
-                                        std::initializer_list<ETags> includeTags = {}) {
-            if (HasTags(includeTags)) {
-                ReportStatus(FindMaxStatus(includeTags), message, setTag, includeTags);
-            }
-        }
-
-        Ydb::Monitoring::StatusFlag::Status GetOverallStatus() const {
-            return OverallStatus;
-        }
-
-        void SetOverallStatus(Ydb::Monitoring::StatusFlag::Status status) {
-            OverallStatus = status;
-        }
-
-        void InheritFrom(TSelfCheckResult& lower) {
-            if (lower.GetOverallStatus() >= OverallStatus) {
-                OverallStatus = lower.GetOverallStatus();
-            }
-            IssueRecords.splice(IssueRecords.end(), std::move(lower.IssueRecords));
-        }
-    };
-
-    struct TSelfCheckContext : TSelfCheckResult {
-        TSelfCheckResult* Upper;
-
-        TSelfCheckContext(TSelfCheckResult* upper)
-            : Upper(upper)
-        {
-            if (Upper) {
-                Location.CopyFrom(upper->Location);
-                Level = upper->Level + 1;
-            }
-        }
-
-        TSelfCheckContext(TSelfCheckResult* upper, const TString& type)
-            : TSelfCheckContext(upper)
-        {
-            Type = type;
-        }
-
-        TSelfCheckContext(const TSelfCheckContext&) = delete;
-
-        ~TSelfCheckContext() {
-            if (Upper) {
-                Upper->InheritFrom(*this);
-            }
-        }
-    };
-
-    template<typename T>
-    struct TRequestResponse {
-        std::variant<std::monostate, std::unique_ptr<T>, TString> Response;
-        NWilson::TSpan Span;
-
-        TRequestResponse() = default;
-        TRequestResponse(NWilson::TSpan&& span)
-            : Span(std::move(span))
-        {}
-
-        TRequestResponse(const TRequestResponse&) = delete;
-        TRequestResponse(TRequestResponse&&) = default;
-        TRequestResponse& operator =(const TRequestResponse&) = delete;
-        TRequestResponse& operator =(TRequestResponse&&) = default;
-
-        void Set(std::unique_ptr<T>&& response) {
-            constexpr bool hasErrorCheck = requires(const std::unique_ptr<T>& r) {TSelfCheckRequest::IsSuccess(r);};
-            if constexpr (hasErrorCheck) {
-                if (!TSelfCheckRequest::IsSuccess(response)) {
-                    Error(TSelfCheckRequest::GetError(response));
-                    return;
-                }
-            }
-            if (!IsDone()) {
-                Span.EndOk();
-            }
-            Response = std::move(response);
-        }
-
-        void Set(TAutoPtr<TEventHandle<T>>&& response) {
-            Set(std::unique_ptr<T>(response->Release().Release()));
-        }
-
-        bool Error(const TString& error) {
-            if (!IsDone()) {
-                Span.EndError(error);
-                Response = error;
-                return true;
-            }
-            return false;
-        }
-
-        bool IsOk() const {
-            return std::holds_alternative<std::unique_ptr<T>>(Response);
-        }
-
-        bool IsError() const {
-            return std::holds_alternative<TString>(Response);
-        }
-
-        bool IsDone() const {
-            return Response.index() != 0;
-        }
-
-        explicit operator bool() const {
-            return IsOk();
-        }
-
-        T* Get() {
-            return std::get<std::unique_ptr<T>>(Response).get();
-        }
-
-        const T* Get() const {
-            return std::get<std::unique_ptr<T>>(Response).get();
-        }
-
-        T& GetRef() {
-            return *Get();
-        }
-
-        const T& GetRef() const {
-            return *Get();
-        }
-
-        T* operator ->() {
-            return Get();
-        }
-
-        const T* operator ->() const {
-            return Get();
-        }
-
-        T& operator *() {
-            return GetRef();
-        }
-
-        const T& operator *() const {
-            return GetRef();
-        }
-
-        TString GetError() const {
-            return std::get<TString>(Response);
-        }
-
-        void Event(const TString& name) {
-            if (Span) {
-                Span.Event(name);
-            }
-        }
-    };
-
-    static bool IsSuccess(const std::unique_ptr<TEvTxProxySchemeCache::TEvNavigateKeySetResult>& ev) {
-        return (ev->Request->ResultSet.size() > 0) && (std::find_if(ev->Request->ResultSet.begin(), ev->Request->ResultSet.end(),
-            [](const auto& entry) {
-                return entry.Status == TSchemeCacheNavigate::EStatus::Ok;
-            }) != ev->Request->ResultSet.end());
-    }
-
-    static TString GetError(const std::unique_ptr<TEvTxProxySchemeCache::TEvNavigateKeySetResult>& ev) {
-        if (ev->Request->ResultSet.size() == 0) {
-            return "empty response";
-        }
-        for (const auto& entry : ev->Request->ResultSet) {
-            if (entry.Status != TSchemeCacheNavigate::EStatus::Ok) {
-                switch (entry.Status) {
-                    case TSchemeCacheNavigate::EStatus::Ok:
-                        return "Ok";
-                    case TSchemeCacheNavigate::EStatus::Unknown:
-                        return "Unknown";
-                    case TSchemeCacheNavigate::EStatus::RootUnknown:
-                        return "RootUnknown";
-                    case TSchemeCacheNavigate::EStatus::PathErrorUnknown:
-                        return "PathErrorUnknown";
-                    case TSchemeCacheNavigate::EStatus::PathNotTable:
-                        return "PathNotTable";
-                    case TSchemeCacheNavigate::EStatus::PathNotPath:
-                        return "PathNotPath";
-                    case TSchemeCacheNavigate::EStatus::TableCreationNotComplete:
-                        return "TableCreationNotComplete";
-                    case TSchemeCacheNavigate::EStatus::LookupError:
-                        return "LookupError";
-                    case TSchemeCacheNavigate::EStatus::RedirectLookupError:
-                        return "RedirectLookupError";
-                    case TSchemeCacheNavigate::EStatus::AccessDenied:
-                        return "AccessDenied";
-                    default:
-                        return ::ToString(static_cast<int>(entry.Status));
-                }
-            }
-        }
-        return "no error";
-    }
-
-    static bool IsSuccess(const std::unique_ptr<TEvSchemeShard::TEvDescribeSchemeResult>& ev) {
-        return ev->GetRecord().status() == NKikimrScheme::StatusSuccess;
-    }
-
-    static TString GetError(const std::unique_ptr<TEvSchemeShard::TEvDescribeSchemeResult>& ev) {
-        return NKikimrScheme::EStatus_Name(ev->GetRecord().status());
-    }
-
-    static bool IsSuccess(const std::unique_ptr<TEvStateStorage::TEvBoardInfo>& ev) {
-        return ev->Status == TEvStateStorage::TEvBoardInfo::EStatus::Ok;
-    }
-
-    static TString GetError(const std::unique_ptr<TEvStateStorage::TEvBoardInfo>& ev) {
-        switch (ev->Status) {
-            case TEvStateStorage::TEvBoardInfo::EStatus::Ok:
-                return "Ok";
-            case TEvStateStorage::TEvBoardInfo::EStatus::Unknown:
-                return "Unknown";
-            case TEvStateStorage::TEvBoardInfo::EStatus::NotAvailable:
-                return "NotAvailable";
-        }
-    }
-
     TString FilterDatabase;
     THashMap<TSubDomainKey, TString> FilterDomainKey;
     TVector<TActorId> PipeClients;
@@ -705,72 +186,9 @@ public:
 
     THashMap<ui64, TRequestResponse<TEvTxProxySchemeCache::TEvNavigateKeySetResult>> NavigateKeySet;
 
-    struct THintOverloadedShard {
-        TTabletId TabletId;
-        TFollowerId FollowerId;
-        double CPUCores;
-        TTabletId SchemeShardId;
-        TString Message;
-    };
-
     THashMap<TString, THintOverloadedShard> OverloadedShardHints;
     static constexpr size_t MAX_OVERLOADED_SHARDS_HINTS = 10;
     static constexpr double OVERLOADED_SHARDS_CPU_CORES = 0.75;
-
-    struct TTabletRequestsState {
-        struct TTabletState {
-            TTabletTypes::EType Type = TTabletTypes::Unknown;
-            TString Database;
-            bool IsUnresponsive = false;
-            TDuration MaxResponseTime;
-            TActorId TabletPipe = {};
-        };
-
-        struct TRequestState {
-            TTabletId TabletId;
-            TString Key;
-            TMonotonic StartTime;
-        };
-
-        std::unordered_map<TTabletId, TTabletState> TabletStates;
-        std::unordered_map<ui64, TRequestState> RequestsInFlight;
-        ui64 RequestId = 0;
-
-        // Some tablets (currently only BSC sys view) do not set response cookie
-        // So we use special constant ids for those requests
-        enum ERequestId : ui64 {
-            RequestStoragePools = 1'000'000,
-            RequestGroups,
-            RequestVSlots,
-            RequestPDisks,
-            RequestGetPartitionStats,
-        };
-
-        void MakeRequest(TTabletId tabletId, const TString& key, ui64 requestId) {
-            RequestsInFlight.emplace(requestId, TRequestState{tabletId, key, TMonotonic::Now()});
-        }
-
-        ui64 MakeRequest(TTabletId tabletId, const TString& key) {
-            MakeRequest(tabletId, key, ++RequestId);
-            return RequestId;
-        }
-
-        TTabletId CompleteRequest(ui64 requestId) {
-            TTabletId tabletId = {};
-            TMonotonic finishTime = TMonotonic::Now();
-            auto itRequest = RequestsInFlight.find(requestId);
-            if (itRequest != RequestsInFlight.end()) {
-                TDuration responseTime = finishTime - itRequest->second.StartTime;
-                tabletId = itRequest->second.TabletId;
-                TTabletState& tabletState = TabletStates[tabletId];
-                if (responseTime > tabletState.MaxResponseTime) {
-                    tabletState.MaxResponseTime = responseTime;
-                }
-                RequestsInFlight.erase(itRequest);
-            }
-            return tabletId;
-        }
-    };
 
     TTabletRequestsState TabletRequests;
 
@@ -1363,14 +781,6 @@ public:
         RequestDone("TEvNodesInfo");
     }
 
-    bool IsStaticGroup(TGroupId groupId) {
-        return !(groupId & 0x80000000);
-    }
-
-    bool NeedWhiteboardForStaticGroupsWithUnknownStatus() {
-        return NodeWardenStorageConfig && !IsSpecificDatabaseFilter();
-    }
-
     void Handle(TEvSysView::TEvGetStoragePoolsResponse::TPtr& ev) {
         TabletRequests.CompleteRequest(TTabletRequestsState::RequestStoragePools);
         StoragePools->Set(std::move(ev));
@@ -1743,18 +1153,6 @@ public:
         }
     }
 
-    static Ydb::Monitoring::StatusFlag::Status MaxStatus(Ydb::Monitoring::StatusFlag::Status a, Ydb::Monitoring::StatusFlag::Status b) {
-        return static_cast<Ydb::Monitoring::StatusFlag::Status>(std::max<int>(a, b));
-    }
-
-    static Ydb::Monitoring::StatusFlag::Status MinStatus(Ydb::Monitoring::StatusFlag::Status a, Ydb::Monitoring::StatusFlag::Status b) {
-        return static_cast<Ydb::Monitoring::StatusFlag::Status>(std::min<int>(a, b));
-    }
-
-    static TString GetNodeLocation(const TEvInterconnect::TNodeInfo& nodeInfo) {
-        return TStringBuilder() << nodeInfo.NodeId << '/' << nodeInfo.Host << ':' << nodeInfo.Port;
-    }
-
     static void Check(TSelfCheckContext& context, const NKikimrWhiteboard::TSystemStateInfo::TPoolStats& poolStats) {
         if (poolStats.name() == "System" || poolStats.name() == "IC" || poolStats.name() == "IO") {
             if (poolStats.usage() >= 0.99) {
@@ -2058,76 +1456,6 @@ public:
         computeStatus.set_overall(context.GetOverallStatus());
     }
 
-    static TString GetVSlotId(const NKikimrSysView::TVSlotKey& vSlotKey) {
-        return TStringBuilder()
-                << vSlotKey.GetNodeId() << '-'
-                << vSlotKey.GetPDiskId() << '-'
-                << vSlotKey.GetVSlotId();
-    }
-
-    static TString GetVDiskId(const NKikimrBlobStorage::TVDiskID& protoVDiskId) {
-        return TStringBuilder()
-                << protoVDiskId.groupid() << '-'
-                << protoVDiskId.groupgeneration() << '-'
-                << protoVDiskId.ring() << '-'
-                << protoVDiskId.domain() << '-'
-                << protoVDiskId.vdisk();
-    }
-
-    static TString GetVDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& protoVSlot) {
-        return TStringBuilder()
-                << protoVSlot.groupid() << '-'
-                << protoVSlot.groupgeneration() << '-'
-                << protoVSlot.failrealmidx() << '-'
-                << protoVSlot.faildomainidx() << '-'
-                << protoVSlot.vdiskidx();
-    }
-
-    static TString GetVDiskId(const NKikimrBlobStorage::TNodeWardenServiceSet_TVDisk& protoVDiskId) {
-        return GetVDiskId(protoVDiskId.vdiskid());
-    }
-
-    static TString GetVDiskId(const NKikimrWhiteboard::TVDiskStateInfo vDiskInfo) {
-        return GetVDiskId(vDiskInfo.vdiskid());
-    }
-
-    static TString GetVDiskId(const NKikimrSysView::TVSlotInfo& vSlot) {
-        return TStringBuilder()
-               << vSlot.GetGroupId() << '-'
-               << vSlot.GetGroupGeneration() << '-'
-               << vSlot.GetFailRealm() << '-'
-               << vSlot.GetFailDomain() << '-'
-               << vSlot.GetVDisk();
-    }
-
-    static TString GetPDiskId(const NKikimrWhiteboard::TVDiskStateInfo vDiskInfo) {
-        return TStringBuilder() << vDiskInfo.nodeid() << "-" << vDiskInfo.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrWhiteboard::TPDiskStateInfo pDiskInfo) {
-        return TStringBuilder() << pDiskInfo.nodeid() << "-" << pDiskInfo.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TPDisk& pDisk) {
-        return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrBlobStorage::TBaseConfig::TVSlot& vSlot) {
-        return TStringBuilder() << vSlot.vslotid().nodeid() << "-" << vSlot.vslotid().pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrBlobStorage::TNodeWardenServiceSet_TPDisk& pDisk) {
-        return TStringBuilder() << pDisk.nodeid() << "-" << pDisk.pdiskid();
-    }
-
-    static TString GetPDiskId(const NKikimrSysView::TVSlotKey& vSlotKey) {
-        return TStringBuilder() << vSlotKey.GetNodeId() << "-" << vSlotKey.GetPDiskId();
-    }
-
-    static TString GetPDiskId(const NKikimrSysView::TPDiskKey& pDiskKey) {
-        return TStringBuilder() << pDiskKey.GetNodeId() << "-" << pDiskKey.GetPDiskId();
-    }
-
     void FillPDiskStatus(const TString& pDiskId, Ydb::Monitoring::StoragePDiskStatus& storagePDiskStatus, TSelfCheckContext context) {
         context.Location.clear_database(); // PDisks are shared between databases
         context.Location.mutable_storage()->mutable_pool()->clear_name(); // PDisks are shared between pools
@@ -2206,21 +1534,6 @@ public:
             case NKikimrBlobStorage::TPDiskSpaceColor::ORANGE:
                 return Ydb::Monitoring::StatusFlag::ORANGE;
             case NKikimrBlobStorage::TPDiskSpaceColor::RED:
-                return Ydb::Monitoring::StatusFlag::RED;
-            default:
-                return Ydb::Monitoring::StatusFlag::UNSPECIFIED;
-        }
-    }
-
-    static Ydb::Monitoring::StatusFlag::Status GetFlagFromWhiteboardFlag(NKikimrWhiteboard::EFlag flag) {
-        switch (flag) {
-            case NKikimrWhiteboard::EFlag::Green:
-                return Ydb::Monitoring::StatusFlag::GREEN;
-            case NKikimrWhiteboard::EFlag::Yellow:
-                return Ydb::Monitoring::StatusFlag::YELLOW;
-            case NKikimrWhiteboard::EFlag::Orange:
-                return Ydb::Monitoring::StatusFlag::ORANGE;
-            case NKikimrWhiteboard::EFlag::Red:
                 return Ydb::Monitoring::StatusFlag::RED;
             default:
                 return Ydb::Monitoring::StatusFlag::UNSPECIFIED;
@@ -2516,82 +1829,6 @@ public:
         storageVDiskStatus.set_overall(context.GetOverallStatus());
     }
 
-    class TGroupChecker {
-        TString ErasureSpecies;
-        bool LayoutCorrect;
-        int FailedDisks = 0;
-        std::array<int, Ydb::Monitoring::StatusFlag::Status_ARRAYSIZE> DisksColors = {};
-        TStackVec<std::pair<ui32, int>> FailedRealms;
-
-        void IncrementFor(ui32 realm) {
-            auto itRealm = FindIf(FailedRealms, [realm](const std::pair<ui32, int>& p) -> bool {
-                return p.first == realm;
-            });
-            if (itRealm == FailedRealms.end()) {
-                itRealm = FailedRealms.insert(FailedRealms.end(), { realm, 1 });
-            } else {
-                itRealm->second++;
-            }
-        }
-
-    public:
-        TGroupChecker(const TString& erasure, const bool layoutCorrect = true)
-            : ErasureSpecies(erasure)
-            , LayoutCorrect(layoutCorrect)
-        {}
-
-        void AddVDiskStatus(Ydb::Monitoring::StatusFlag::Status status, ui32 realm) {
-            ++DisksColors[status];
-            switch (status) {
-                case Ydb::Monitoring::StatusFlag::BLUE: // disk is good, but not available
-                // No yellow or orange status here - this is intentional - they are used when a disk is running out of space, but is currently available
-                case Ydb::Monitoring::StatusFlag::RED: // disk is bad, probably not available
-                case Ydb::Monitoring::StatusFlag::GREY: // the status is absent, the disk is not available
-                    IncrementFor(realm);
-                    ++FailedDisks;
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        void ReportStatus(TSelfCheckContext& context) const {
-            context.OverallStatus = Ydb::Monitoring::StatusFlag::GREEN;
-            if (!LayoutCorrect) {
-                context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Group layout is incorrect", ETags::GroupState);
-            }
-            if (ErasureSpecies == NONE) {
-                if (FailedDisks > 0) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Group failed", ETags::GroupState, {ETags::VDiskState});
-                }
-            } else if (ErasureSpecies == BLOCK_4_2) {
-                if (FailedDisks > 2) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Group failed", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedDisks > 1) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Group has no redundancy", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedDisks > 0) {
-                    if (DisksColors[Ydb::Monitoring::StatusFlag::BLUE] == FailedDisks) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    } else {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    }
-                }
-            } else if (ErasureSpecies == MIRROR_3_DC) {
-                if (FailedRealms.size() > 2 || (FailedRealms.size() == 2 && FailedRealms[0].second > 1 && FailedRealms[1].second > 1)) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::RED, "Group failed", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedRealms.size() == 2) {
-                    context.ReportStatus(Ydb::Monitoring::StatusFlag::ORANGE, "Group has no redundancy", ETags::GroupState, {ETags::VDiskState});
-                } else if (FailedDisks > 0) {
-                    if (DisksColors[Ydb::Monitoring::StatusFlag::BLUE] == FailedDisks) {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::BLUE, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    } else {
-                        context.ReportStatus(Ydb::Monitoring::StatusFlag::YELLOW, "Group degraded", ETags::GroupState, {ETags::VDiskState});
-                    }
-                }
-            }
-        }
-    };
-
     void FillGroupStatusWithWhiteboard(TGroupId groupId, const NKikimrWhiteboard::TBSGroupStateInfo& groupInfo, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckContext context) {
         if (context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_id()->empty()) {
             context.Location.mutable_storage()->mutable_pool()->mutable_group()->add_id();
@@ -2633,263 +1870,6 @@ public:
         storageGroupStatus.set_overall(context.GetOverallStatus());
     }
 
-    static const inline TString NONE = "none";
-    static const inline TString BLOCK_4_2 = "block-4-2";
-    static const inline TString MIRROR_3_DC = "mirror-3-dc";
-    static const int MERGING_IGNORE_SIZE = 4;
-
-    struct TMergeIssuesContext {
-        std::unordered_map<ETags, TList<TSelfCheckContext::TIssueRecord>> recordsMap;
-        std::unordered_set<TString> removeIssuesIds;
-
-        TMergeIssuesContext(TList<TSelfCheckContext::TIssueRecord>& records) {
-            for (auto it = records.begin(); it != records.end(); ) {
-                auto move = it++;
-                recordsMap[move->Tag].splice(recordsMap[move->Tag].end(), records, move);
-            }
-        }
-
-        void RemoveUnlinkIssues(TList<TSelfCheckContext::TIssueRecord>& records) {
-            bool isRemovingIssuesIteration = true;
-            while (isRemovingIssuesIteration) {
-                isRemovingIssuesIteration = false;
-
-                std::unordered_set<TString> necessaryIssuesIds;
-                for (auto it = records.begin(); it != records.end(); it++) {
-                    auto reasons = it->IssueLog.reason();
-                    for (auto reasonIt = reasons.begin(); reasonIt != reasons.end(); reasonIt++) {
-                        necessaryIssuesIds.insert(*reasonIt);
-                    }
-                }
-
-                for (auto it = records.begin(); it != records.end(); ) {
-                    if (!necessaryIssuesIds.contains(it->IssueLog.id()) && removeIssuesIds.contains(it->IssueLog.id())) {
-                        auto reasons = it->IssueLog.reason();
-                        for (auto reasonIt = reasons.begin(); reasonIt != reasons.end(); reasonIt++) {
-                            removeIssuesIds.insert(*reasonIt);
-                        }
-                        isRemovingIssuesIteration = true;
-                        it = records.erase(it);
-                    } else {
-                        it++;
-                    }
-                }
-            }
-
-            {
-                std::unordered_set<TString> issueIds;
-                for (auto it = records.begin(); it != records.end(); it++) {
-                    issueIds.insert(it->IssueLog.id());
-                }
-
-                for (auto it = records.begin(); it != records.end(); it++) {
-                    auto reasons = it->IssueLog.mutable_reason();
-                    for (auto reasonIt = reasons->begin(); reasonIt != reasons->end(); ) {
-                        if (!issueIds.contains(*reasonIt)) {
-                            reasonIt = reasons->erase(reasonIt);
-                        } else {
-                            reasonIt++;
-                        }
-                    }
-                }
-            }
-        }
-
-        void RenameMergingIssues(TList<TSelfCheckContext::TIssueRecord>& records) {
-            for (auto it = records.begin(); it != records.end(); it++) {
-                if (it->IssueLog.count() > 0) {
-                    TString message = it->IssueLog.message();
-                    switch (it->Tag) {
-                        case ETags::GroupState: {
-                            message = std::regex_replace(message.c_str(), std::regex("^Group has "), "Groups have ");
-                            message = std::regex_replace(message.c_str(), std::regex("^Group is "), "Groups are ");
-                            message = std::regex_replace(message.c_str(), std::regex("^Group "), "Groups ");
-                            break;
-                        }
-                        case ETags::VDiskState: {
-                            message = std::regex_replace(message.c_str(), std::regex("^VDisk has "), "VDisks have ");
-                            message = std::regex_replace(message.c_str(), std::regex("^VDisk is "), "VDisks are ");
-                            message = std::regex_replace(message.c_str(), std::regex("^VDisk "), "VDisks ");
-                            break;
-                        }
-                        case ETags::PDiskState: {
-                            message = std::regex_replace(message.c_str(), std::regex("^PDisk has "), "PDisks have ");
-                            message = std::regex_replace(message.c_str(), std::regex("^PDisk is "), "PDisks are ");
-                            message = std::regex_replace(message.c_str(), std::regex("^PDisk "), "PDisks ");
-                            break;
-                        }
-                        default:
-                            break;
-                    }
-                    it->IssueLog.set_message(message);
-                }
-            }
-        }
-
-        void FillRecords(TList<TSelfCheckContext::TIssueRecord>& records) {
-            for(auto it = recordsMap.begin(); it != recordsMap.end(); ++it) {
-                records.splice(records.end(), it->second);
-            }
-            RemoveUnlinkIssues(records);
-            RenameMergingIssues(records);
-        }
-
-        TList<TSelfCheckContext::TIssueRecord>& GetRecords(ETags tag) {
-            return recordsMap[tag];
-        }
-    };
-
-    bool FindRecordsForMerge(TList<TSelfCheckContext::TIssueRecord>& records, TList<TSelfCheckContext::TIssueRecord>& similar, TList<TSelfCheckContext::TIssueRecord>& merged) {
-        while (!records.empty() && similar.empty()) {
-            similar.splice(similar.end(), records, records.begin());
-            for (auto it = records.begin(); it != records.end(); ) {
-                bool isSimilar = it->IssueLog.status() == similar.begin()->IssueLog.status()
-                    && it->IssueLog.message() == similar.begin()->IssueLog.message()
-                    && it->IssueLog.level() == similar.begin()->IssueLog.level() ;
-                if (isSimilar && similar.begin()->Tag == ETags::VDiskState) {
-                    isSimilar = it->IssueLog.location().storage().node().id() == similar.begin()->IssueLog.location().storage().node().id();
-                }
-                if (isSimilar) {
-                    auto move = it++;
-                    similar.splice(similar.end(), records, move);
-                } else {
-                    ++it;
-                }
-            }
-
-            if (similar.size() == 1) {
-                merged.splice(merged.end(), similar);
-            }
-        }
-
-        return !similar.empty();
-    }
-
-    std::shared_ptr<TList<TSelfCheckContext::TIssueRecord>> FindChildrenRecords(TList<TSelfCheckContext::TIssueRecord>& records, TSelfCheckContext::TIssueRecord& parent) {
-        std::shared_ptr<TList<TSelfCheckContext::TIssueRecord>> children(new TList<TSelfCheckContext::TIssueRecord>);
-        std::unordered_set<TString> childrenIds;
-        for (auto reason: parent.IssueLog.reason()) {
-            childrenIds.insert(reason);
-        }
-
-        for (auto it = records.begin(); it != records.end(); ) {
-            if (childrenIds.contains(it->IssueLog.id())) {
-                auto move = it++;
-                children->splice(children->end(), records, move);
-            } else {
-                it++;
-            }
-        }
-
-        return children;
-    }
-
-    void MoveDataInFirstRecord(TMergeIssuesContext& context, TList<TSelfCheckContext::TIssueRecord>& similar) {
-        auto mainReasons = similar.begin()->IssueLog.mutable_reason();
-        std::unordered_set<TString> ids;
-        ids.insert(similar.begin()->IssueLog.id());
-        std::unordered_set<TString> mainReasonIds;
-        for (auto it = mainReasons->begin(); it != mainReasons->end(); it++) {
-            mainReasonIds.insert(*it);
-        }
-
-        for (auto it = std::next(similar.begin(), 1); it != similar.end(); ) {
-            if (ids.contains(it->IssueLog.id())) {
-                it++;
-                continue;
-            }
-            ids.insert(it->IssueLog.id());
-
-            switch (similar.begin()->Tag) {
-                case ETags::GroupState: {
-                    auto mainGroupIds = similar.begin()->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_id();
-                    auto donorGroupIds = it->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_id();
-                    mainGroupIds->Add(donorGroupIds->begin(), donorGroupIds->end());
-                    break;
-                }
-                case ETags::VDiskState: {
-                    auto mainVdiskIds = similar.begin()->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id();
-                    auto donorVdiskIds = it->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_id();
-                    mainVdiskIds->Add(donorVdiskIds->begin(), donorVdiskIds->end());
-                    break;
-                }
-                case ETags::PDiskState: {
-                    auto mainPdisk = similar.begin()->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk();
-                    auto donorPdisk = it->IssueLog.mutable_location()->mutable_storage()->mutable_pool()->mutable_group()->mutable_vdisk()->mutable_pdisk();
-                    mainPdisk->Add(donorPdisk->begin(), donorPdisk->end());
-                    break;
-                }
-                default:
-                    break;
-            }
-
-            auto donorReasons = it->IssueLog.mutable_reason();
-            for (auto donorReasonIt = donorReasons->begin(); donorReasonIt != donorReasons->end(); donorReasonIt++) {
-                if (!mainReasonIds.contains(*donorReasonIt)) {
-                    mainReasons->Add(donorReasonIt->c_str());
-                    mainReasonIds.insert(*donorReasonIt);
-                }
-            }
-
-            context.removeIssuesIds.insert(it->IssueLog.id());
-            it = similar.erase(it);
-        }
-
-        similar.begin()->IssueLog.set_count(ids.size());
-        similar.begin()->IssueLog.set_listed(ids.size());
-    }
-
-    void MergeLevelRecords(TMergeIssuesContext& context, TList<TSelfCheckContext::TIssueRecord>& records) {
-        TList<TSelfCheckContext::TIssueRecord> handled;
-        while (!records.empty()) {
-            TList<TSelfCheckContext::TIssueRecord> similar;
-            if (FindRecordsForMerge(records, similar, handled)) {
-                MoveDataInFirstRecord(context, similar);
-                handled.splice(handled.end(), similar, similar.begin());
-            }
-        }
-        records.splice(records.end(), handled);
-    }
-
-    void MergeLevelRecords(TMergeIssuesContext& context, ETags levelTag) {
-        auto& records = context.GetRecords(levelTag);
-        MergeLevelRecords(context, records);
-    }
-
-    void MergeLevelRecords(TMergeIssuesContext& context, ETags levelTag, ETags upperTag) {
-        auto& levelRecords = context.GetRecords(levelTag);
-        auto& upperRecords = context.GetRecords(upperTag);
-
-        for (auto it = upperRecords.begin(); it != upperRecords.end(); it++) {
-            auto children = FindChildrenRecords(levelRecords, *it);
-            if (children->size() > 1) {
-                MergeLevelRecords(context, *children);
-            }
-            levelRecords.splice(levelRecords.end(), *children);
-        }
-    }
-
-    int GetIssueCount(TSelfCheckContext::TIssueRecord& record) {
-        return record.IssueLog.count() == 0 ? 1 : record.IssueLog.count();
-    }
-
-    void SetIssueCount(TSelfCheckContext::TIssueRecord& record, int value) {
-        if (record.IssueLog.listed() == 0) {
-            record.IssueLog.set_listed(1);
-        }
-        record.IssueLog.set_count(value);
-    }
-
-    int GetIssueListed(TSelfCheckContext::TIssueRecord& record) {
-        return record.IssueLog.listed() == 0 ? 1 : record.IssueLog.listed();
-    }
-
-    void SetIssueListed(TSelfCheckContext::TIssueRecord& record, int value) {
-        if (record.IssueLog.count() == 0) {
-            record.IssueLog.set_count(1);
-        }
-        record.IssueLog.set_listed(value);
-    }
 
     void FillGroupStatus(TGroupId groupId, Ydb::Monitoring::StorageGroupStatus& storageGroupStatus, TSelfCheckContext context) {
         context.Location.mutable_storage()->mutable_pool()->mutable_group()->mutable_id()->Clear();
@@ -2927,9 +1907,9 @@ public:
     void MergeRecords(TList<TSelfCheckContext::TIssueRecord>& records) {
         TMergeIssuesContext mergeContext(records);
         if (Request->Request.merge_records()) {
-            MergeLevelRecords(mergeContext, ETags::GroupState);
-            MergeLevelRecords(mergeContext, ETags::VDiskState, ETags::GroupState);
-            MergeLevelRecords(mergeContext, ETags::PDiskState, ETags::VDiskState);
+            mergeContext.MergeLevelRecords(ETags::GroupState);
+            mergeContext.MergeLevelRecords(ETags::VDiskState, ETags::GroupState);
+            mergeContext.MergeLevelRecords(ETags::PDiskState, ETags::VDiskState);
         }
         mergeContext.FillRecords(records);
     }
